chore: release binaries
