+++ npm/darwin-arm64/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "supabase-darwin-arm64",
+  "version": "0.0.0-automated",
+  "license": "MIT",
+  "os": [
+    "darwin"
+  ],
+  "cpu": [
+    "arm64"
+  ],
+  "files": [
+    "supabase"
+  ]
+}
