+++ npm/windows-arm64/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "supabase-windows-arm64",
+  "version": "0.0.0-automated",
+  "license": "MIT",
+  "os": [
+    "win32"
+  ],
+  "cpu": [
+    "arm64"
+  ],
+  "files": [
+    "supabase.exe"
+  ]
+}
